Déplacements horizontaux de Goomba. Sans collisions

+++ core/src/com/smeshed/mb/Utils/Coordonnees.java
@@ -1,27 +1,27 @@
 package com.smeshed.mb.Utils;
 
 public class Coordonnees {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
-    public Coordonnees(float x, float y) {
+    public Coordonnees(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
